+++ CHANGELOG.md
@@ -33,6 +33,14 @@
 
 
 ### Changed
+
+* The cache now stores desktop file **IDs** instead of the application
+  titles. This way, we do not store multiple cache entries with the
+  same value (title) ([#339][339]).
+
+[339]: https://codeberg.org/dnkl/fuzzel/issues/339
+
+
 ### Deprecated
 ### Removed
 ### Fixed